Moved RateFormat into separate file
